[DAD-598] refactor: scrap content 타입 분리
