+++ src/components/molcules/VideoScrapCard.tsx
@@ -1,6 +1,6 @@
 import styled from 'styled-components';
 
-import { contentProps } from '../organisms/ScrapCard';
+import { contentProps } from '../../types/ContentType';
 import Memo from './Memo';
 import Button from '../atoms/DefaultButton';
 import Chip from '../atoms/Chip';
